+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Improve resource card layout [#65](https://github.com/IN-CORE/incore-studio/issues/65)
 - Disable unused buttons from navbar [#215](https://github.com/IN-CORE/incore-studio/issues/215)
+- Project Dashboard look and feel [#216](https://github.com/IN-CORE/incore-studio/issues/216)
 
 ## [Beta-1] - 05-01-2025